Audio/DrvAudio: Windows build fix, size_t/uint32_t warning
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@73576 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Devices/Audio/DrvAudio.cpp
@@ -1600,7 +1600,7 @@ static int drvAudioStreamCaptureNonInterleaved(PDRVAUDIO pThis, PPDMAUDIOSTREAM 
     {
         uint32_t cbCaptured;
         rc = pThis->pHostDrvAudio->pfnStreamCapture(pThis->pHostDrvAudio, pStream->pvBackend,
-                                                    auBuf, RT_MIN(cbReadable, cbBuf), &cbCaptured);
+                                                    auBuf, RT_MIN(cbReadable, (uint32_t)cbBuf), &cbCaptured);
         if (RT_FAILURE(rc))
         {
             int rc2 = drvAudioStreamControlInternalBackend(pThis, pStream, PDMAUDIOSTREAMCMD_DISABLE);
